text_speech_api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
boto3==1.35.72
httpx==0.28.0
python-multipart==0.0.20
pillow==11.0.0
pyjwt[crypto]==2.10.1
python-dotenv==1.0.1
gTTS==2.5.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
